+++ src/main/java/ua/karazin/interfaces/ProjectLibrary/dto/BooksInfoDTO.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 public record BooksInfoDTO(
         @NotNull
-        Integer isbn,
+        Long isbn,
 
         @NotBlank(message = "Title is required")
         String title,
